Create randomNums.py
I had some doubts about just how random randint was, so i created this to see just how often each number is pulled. As the number of cycles increases you'll see that we average out around 20% for each number pulled, which is exactly what you'd expect to see if the numbers were random, so as far as random goes, sudo random is perfect for my needs.